refactor(runtime): extract daemon phase orchestration module (#3735)
